Puredata MIDI droplet visualizer
This patch is designed for piano, it takes incoming MIDI notes and generates a visualization based on the notes played. Incoming MIDI notes are placed on a circle of fifths and colour is assigned to their position on the piano. Brightness and rate of growth are determined by velocity.

It's a little glitchy and the design is not particularly elegant but it makes for an interesting visual representation of the circle of fifths.